Implement Role-Based Authorization for Brand Endpoints, also Added Dashbaoard in the front
Added role-based authorization for brand endpoints to restrict access based on admin role. Added Dashboard page only for admin. Also implemented a admin sidebar in the dashboard.

+++ src/Modules/Catalog/Catalog/Brands/Features/DeleteBrand/DeleteBrandEndpoint.cs
@@ -1,13 +1,19 @@
-﻿namespace Catalog.Brands.Features.DeleteBrand
+﻿using System.Security.Claims;
+
+namespace Catalog.Brands.Features.DeleteBrand
 {
     public class DeleteBrandEndpoint : ICarterModule
     {
         public void AddRoutes(IEndpointRouteBuilder app)
         {
-            app.MapDelete("/brands/{id}", async (Guid id, ISender sender) =>
+            app.MapDelete("/brands/{id}", async (Guid id, ISender sender, ClaimsPrincipal user) =>
             {
-                var result = await sender.Send(new DeleteBrandCommand(id));
+                if (!user.IsInRole("admin"))
+                {
+                    return Results.Forbid();
+                }
 
+                var result = await sender.Send(new DeleteBrandCommand(id));
                 var response = new DeleteBrandResult(result.IsSuccessful);
 
                 return Results.Ok(response);
@@ -15,8 +21,9 @@
             .WithName("Delete Brand")
             .Produces<DeleteBrandResult>(StatusCodes.Status200OK)
             .ProducesProblem(StatusCodes.Status404NotFound)
+            .Produces(StatusCodes.Status403Forbidden) 
             .WithSummary("Delete Brand")
-            .WithDescription("Delete a brand by its ID.");
+            .WithDescription("Deletes a brand by its ID. Only accessible by admin.");
         }
     }
 }
